refactor: Move tests to main test files
